fix bugs, merge to main for presentation
